Polish InterceptMethodsBeanDefinitionDecorator
Issue gh-11328
